Some reorganizing, add basic authenticated get
