+++ src/_services/auth-header.js
@@ -1,9 +1,8 @@
 export default function authHeader() {
   let user = JSON.parse(localStorage.getItem('user'));
 
-  if (user && user.accessToken) {
-    return { Authorization: 'Bearer ' + user.accessToken }; // for Spring Boot back-end
-    // return { 'x-access-token': user.accessToken };       // for Node.js Express back-end
+  if (user && user.token) {
+    return { Authorization: 'Bearer ' + user.token };
   } else {
     return {};
   }
