+++ src/router.js
@@ -1,22 +1,22 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Login from './views/Login.vue';
-import Register from './views/Register.vue';
-import ShopItems from "./components/ShopItems";
-import Orders from "./components/Orders";
-import OrderDetail from "./components/OrderDetail";
-import OrderEdit from "./components/OrderEdit";
+import Login from './views/user/Login.vue';
+import Register from './views/user/Register.vue';
+import ProductList from "./views/product/ProductList";
+import OrderList from "./views/order/OrderList";
+import OrderDetail from "./views/order/OrderDetail";
+import OrderEdit from "./views/order/OrderEdit";
 
 Vue.use(Router);
 
 export const router = new Router({
   mode: 'history',
   routes: [
-    { path: '/', name: 'home', component: ShopItems },
+    { path: '/', name: 'home', component: ProductList },
     { path: '/login', component: Login },
     { path: '/register', component: Register },
-    { path: '/profile', name: 'profile', component: () => import('./views/Profile.vue') }, // lazy loaded
-    { path: '/orders', component: Orders },
+    { path: '/profile', name: 'profile', component: () => import('./views/user/Profile.vue') }, // lazy loaded
+    { path: '/orders', component: OrderList },
     { path: '/orders/:id', component: OrderDetail },
     { path: '/orders/:id/edit', component: OrderEdit}
   ]
