+++ src/views/user/Login.vue
@@ -52,7 +52,7 @@
 </template>
 
 <script>
-import User from '../_models/user';
+import User from '../../_models/user';
 
 export default {
   name: 'Login',
